Improve the definition of a "used" require
For requires, we weren't considering whether anything actually referred
to the original name. That is, in (:require [x :as y]) or similar, we
only looked for references to y, not x. If the body didn't refer to y
but did use x, we would erroneously remove the whole require.

Additionally, if the require were simply [x], we would conservatively
keep it.

Fix this by checking for references to x and use that to decide if the
require is used.

Fixes #57.